Fix clang-tidy warnings

+++ src/node.c
@@ -640,8 +640,8 @@ SerdNode*
 serd_new_integer(int64_t i, const SerdNode* datatype)
 {
 	const SerdNode* type      = datatype ? datatype : &serd_xsd_integer.node;
-	int64_t         abs_i     = (i < 0) ? -i : i;
-	const unsigned  digits    = serd_digits((double)abs_i);
+    uint64_t        abs_i     = (uint64_t)((i < 0) ? -i : i);
+    const unsigned  digits    = (unsigned)serd_digits((double)abs_i);
 	const size_t    type_len  = serd_node_total_size(type);
 	const size_t    total_len = digits + 2 + type_len;
